+++ asconfig.json
@@ -1,28 +1,26 @@
 {
-  "options": {
-    "outFile": "build/trip.wasm",
-    "exports": [
-      "memory",
-      "reset",
-      "allocTerminal",
-      "allocCons",
-      "arenaKernelStep",
-      "reduce",
-      "kindOf",
-      "symOf",
-      "leftOf",
-      "rightOf"
-    ],
-    "exportRuntime": false
-  },
   "targets": {
     "debug": {
+      "outFile": "build/debug.wasm",
+      "textFile": "build/debug.wat",
       "sourceMap": true,
       "debug": true
     },
     "release": {
+      "outFile": "build/release.wasm",
+      "textFile": "build/release.wat",
+      "sourceMap": true,
       "optimizeLevel": 3,
-      "shrinkLevel": 0
+      "shrinkLevel": 1,
+      "converge": false,
+      "noAssert": false
     }
+  },
+  "options": {
+    "bindings": "esm",
+    "runtime": "minimal",
+    "importMemory": true,
+    "initialMemory": 512,
+    "maximumMemory": 65536
   }
 }
